+++ Implementacion/src/main/java/dds/grupo4/tpimpacto/services/BaseService.java
@@ -9,5 +9,5 @@ public interface BaseService<T extends BaseEntity> {
 
     List<T> getAll();
 
-    T getById(int id);
+    T getById(long id);
 }
